backend/requirements.txt
fastapi==0.82.0
Pillow~=9.2.0
numpy~=1.23.2
uvicorn~=0.18.3
keras~=2.10.0
tensorflow-cpu~=2.10.0
python-multipart
redis[hiredis]
pydantic[dotenv]